moved dragEndHandler to ItemArea
